// src/api/models/user.ts
/**
 * Generated by orval v7.16.0 🍺
 * Do not edit manually.
 * Aris-blog
 * OpenAPI spec version: 1.0
 */

export interface User {
  /** URL or path to the user's avatar image */
  avatar: string;
  /** Timestamp when the user account was created */
  createdAt?: string;
  /** Email address of the user */
  email?: string;
  /** Timestamp of the user's last login */
  lastLogin?: string;
  /** Display name of the user */
  name: string;
  /** Permission level of the user */
  permission?: string;
  /**
   * Unique identifier for the user
   * @minimum 0
   */
  userID: number;
}